refactor: Update group chat message service to use int64 for IDs
Updated the GroupChatMessageService to use int64 for group and user IDs in its methods. This change ensures consistency with the updated models and database schema, preventing potential data type mismatch errors.

+++ backend/internal/service/group_chat_message_service.go
@@ -17,13 +17,13 @@ func NewGroupChatMessageService(groupChatMessageStore store.GroupChatMessageStor
 	return &groupChatMessageService{groupChatMessageStore: groupChatMessageStore, groupService: groupService, groupMemberStore: groupMemberStore}
 }
 
-func (s *groupChatMessageService) SendGroupChatMessage(groupID, senderID int, content string) (*models.GroupChatMessage, error) {
-	_, err := s.groupService.GetGroupByID(groupID)
+func (s *groupChatMessageService) SendGroupChatMessage(groupID, senderID int64, content string) (*models.GroupChatMessage, error) {
+	_, err := s.groupService.GetGroupByID(int64(groupID))
 	if err != nil {
 		return nil, fmt.Errorf("group not found: %w", err)
 	}
 
-	isMember, err := s.groupMemberStore.IsGroupMember(groupID, senderID)
+	isMember, err := s.groupMemberStore.IsGroupMember(int64(groupID), int64(senderID))
 	if err != nil {
 		return nil, fmt.Errorf("failed to check group membership: %w", err)
 	}
@@ -45,8 +45,8 @@ func (s *groupChatMessageService) SendGroupChatMessage(groupID, senderID int, co
 	return createdMessage, nil
 }
 
-func (s *groupChatMessageService) GetGroupChatMessages(groupID int, userID int, limit, offset int) ([]*models.GroupChatMessage, error) {
-	_, err := s.groupService.GetGroupByID(groupID)
+func (s *groupChatMessageService) GetGroupChatMessages(groupID int64, userID int64, limit, offset int) ([]*models.GroupChatMessage, error) {
+	_, err := s.groupService.GetGroupByID(int64(groupID))
 	if err != nil {
 		return nil, fmt.Errorf("group not found: %w", err)
 	}
